refactor(gateway): own health snapshot collection
