+++ src/commands/health.ts
@@ -1,27 +1,13 @@
-import { expectDefined } from "@openclaw/normalization-core";
 /** Collects and renders gateway health for channels, agents, plugins, and sessions. */
-import { resolveTimerTimeoutMs } from "@openclaw/normalization-core/number-coercion";
 import { asNullableRecord } from "@openclaw/normalization-core/record-coerce";
 import { styleHealthChannelLine } from "../../packages/terminal-core/src/health-style.js";
 import { isRich } from "../../packages/terminal-core/src/theme.js";
-import { listAgentEntries, resolveDefaultAgentId } from "../agents/agent-scope.js";
-import { inspectChannelAccount } from "../channels/account-inspection.js";
-import { redactChannelStatusSummaryBaseUrl } from "../channels/account-snapshot-fields.js";
-import {
-  resolveChannelAccountConfigured,
-  resolveChannelAccountEnabled,
-} from "../channels/account-summary.js";
-import { countFailedChannelIngressQueueEntries } from "../channels/message/ingress-queue.js";
 import { resolveChannelDefaultAccountId } from "../channels/plugins/helpers.js";
 import { listReadOnlyChannelPluginsForConfig } from "../channels/plugins/read-only.js";
-import { buildChannelAccountSnapshotFromAccount } from "../channels/plugins/status.js";
-import type { ChannelPlugin } from "../channels/plugins/types.plugin.js";
-import type { ChannelAccountSnapshot } from "../channels/plugins/types.public.js";
 import { probeGatewayStatus } from "../cli/daemon-cli/probe.js";
 import { withProgress } from "../cli/progress.js";
 import { resolveStorePath } from "../config/sessions/paths.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { listContextEngineQuarantines } from "../context-engine/registry.js";
 import {
   buildGatewayConnectionDetails,
   buildGatewayProbeConnectionDetails,
@@ -29,39 +15,20 @@ import {
   formatGatewayTransportErrorJson,
   isGatewayCredentialsRequiredError,
 } from "../gateway/call.js";
-import {
-  DEFAULT_CHANNEL_CONNECT_GRACE_MS,
-  DEFAULT_CHANNEL_STALE_EVENT_THRESHOLD_MS,
-  evaluateChannelHealth,
-} from "../gateway/channel-health-policy.js";
-import type { GatewayHotReloadStatus } from "../gateway/config-reload-status.types.js";
 import { isGatewaySecretRefUnavailableError } from "../gateway/credentials.js";
-import type {
-  AgentHealthSummary,
-  ChannelAccountHealthSummary,
-  ChannelHealthSummary,
-  ContextEngineHealthSummary,
-  DeliveryQueueHealthSummary,
-  HealthSummary,
-  PluginHealthErrorSummary,
-  PluginHealthSummary,
-} from "../gateway/health/types.js";
-import type { ChannelRuntimeSnapshot } from "../gateway/server-channel-runtime.types.js";
+import { resolveHealthAccountContext } from "../gateway/health/account-context.js";
+import {
+  buildHealthSessionSummary as buildSessionSummary,
+  resolveHealthAgentOrder as resolveAgentOrder,
+} from "../gateway/health/collector.js";
+import type { AgentHealthSummary, HealthSummary } from "../gateway/health/types.js";
 import { info } from "../globals.js";
-import { countFailedDeliveryQueueEntries } from "../infra/delivery-queue-sqlite.js";
 import { isDiagnosticFlagEnabled } from "../infra/diagnostic-flags.js";
 import { formatErrorMessage } from "../infra/errors.js";
 import { formatDurationHuman } from "../infra/format-time/format-duration.js";
 import { resolveHeartbeatSummaryForAgent } from "../infra/heartbeat-summary.js";
 import { createSubsystemLogger } from "../logging/subsystem.js";
-import {
-  degradedPluginMatchesRoot,
-  listActiveDegradedPlugins,
-  toPublicPluginVerificationDiagnostic,
-} from "../plugins/runtime-degraded-state.js";
-import { getActivePluginRegistry } from "../plugins/runtime.js";
 import { buildChannelAccountBindings, resolvePreferredAccountId } from "../routing/bindings.js";
-import { normalizeAgentId } from "../routing/session-key.js";
 import { type RuntimeEnv, writeRuntimeJson } from "../runtime.js";
 import {
   buildCredentialsRequiredHealthDiagnostic,
@@ -136,45 +103,6 @@ export async function emitReachableGatewayAuthDiagnostic(params: {
 
 const loadConfigRuntime = async () => await import("../config/config.js");
 
-const PUBLIC_IMESSAGE_FULL_DISK_ACCESS_ERROR =
-  "imsg cannot access ~/Library/Messages/chat.db. Grant Full Disk Access to the Gateway/launcher process and restart Gateway.";
-
-const redactIMessageProbeErrorMessage = (message: string): string => {
-  const trimmed = message.trim();
-  if (!trimmed) {
-    return "";
-  }
-  return trimmed.replaceAll(
-    /\/Users\/[^/\s]+\/Library\/Messages\/chat\.db/g,
-    "~/Library/Messages/chat.db",
-  );
-};
-
-const buildNonSensitiveProbeFailure = (
-  channelId: string,
-  probe: unknown,
-): Record<string, unknown> | undefined => {
-  const record = asNullableRecord(probe);
-  if (channelId !== "imessage" || !record || record.ok !== false) {
-    return undefined;
-  }
-  if (typeof record.error !== "string") {
-    return undefined;
-  }
-
-  // Preserve the actionable Full Disk Access failure while stripping the local
-  // username path before health leaves the gateway.
-  const error = redactIMessageProbeErrorMessage(record.error);
-  if (
-    !/\bimsg\b/i.test(error) ||
-    !error.includes("~/Library/Messages/chat.db") ||
-    !/\bFull Disk Access\b/i.test(error)
-  ) {
-    return undefined;
-  }
-  return { ok: false, error: PUBLIC_IMESSAGE_FULL_DISK_ACCESS_ERROR };
-};
-
 const formatDurationParts = (ms: number): string => {
   if (!Number.isFinite(ms)) {
     return "unknown";
@@ -218,23 +146,6 @@ function formatEventLoopHealthLine(summary: HealthSummary): string | null {
   }`;
 }
 
-function buildContextEngineHealthSummary(): ContextEngineHealthSummary | undefined {
-  const quarantined: ContextEngineHealthSummary["quarantined"] = [];
-  for (const entry of listContextEngineQuarantines()) {
-    const summary: ContextEngineHealthSummary["quarantined"][number] = {
-      engineId: entry.engineId,
-      operation: entry.operation,
-      reason: entry.reason,
-      failedAt: entry.failedAt.getTime(),
-    };
-    if (entry.owner) {
-      summary.owner = entry.owner;
-    }
-    quarantined.push(summary);
-  }
-  return quarantined.length > 0 ? { quarantined } : undefined;
-}
-
 /** Formats context engine quarantine state for text health output. */
 export function formatContextEngineHealthLine(summary: HealthSummary): string | null {
   const quarantined = summary.contextEngines?.quarantined ?? [];
@@ -243,54 +154,6 @@ export function formatContextEngineHealthLine(summary: HealthSummary): string | 
   }
   const engines = quarantined.map((entry) => entry.engineId).join(", ");
   return `Context engine: warning (${quarantined.length} quarantined; downgraded to legacy: ${engines})`;
-}
-
-/** Builds dead-lettered inbound and outbound queue health for cached gateway responses. */
-export function buildDeliveryQueueHealthSummary(): DeliveryQueueHealthSummary | undefined {
-  // Queue health reads are diagnostic; a storage failure must not take the
-  // gateway health endpoint down with it.
-  let failed: DeliveryQueueHealthSummary["failed"] = [];
-  try {
-    failed = countFailedDeliveryQueueEntries().map((queue) => {
-      const entry: DeliveryQueueHealthSummary["failed"][number] = {
-        queueName: queue.queueName,
-        count: queue.count,
-      };
-      if (queue.oldestFailedAt != null) {
-        entry.oldestFailedAt = queue.oldestFailedAt;
-      }
-      return entry;
-    });
-  } catch (error) {
-    debugHealth(undefined, "outbound delivery queue health read failed", {
-      error: formatErrorMessage(error),
-    });
-  }
-  let ingressFailed: NonNullable<DeliveryQueueHealthSummary["ingressFailed"]> = [];
-  try {
-    ingressFailed = countFailedChannelIngressQueueEntries().map((queue) => {
-      const entry: NonNullable<DeliveryQueueHealthSummary["ingressFailed"]>[number] = {
-        channelId: queue.channelId,
-        accountId: queue.accountId,
-        count: queue.count,
-      };
-      if (queue.oldestFailedAt != null) {
-        entry.oldestFailedAt = queue.oldestFailedAt;
-      }
-      return entry;
-    });
-  } catch (error) {
-    debugHealth(undefined, "channel ingress queue health read failed", {
-      error: formatErrorMessage(error),
-    });
-  }
-  if (failed.length === 0 && ingressFailed.length === 0) {
-    return undefined;
-  }
-  return {
-    failed,
-    ...(ingressFailed.length > 0 ? { ingressFailed } : {}),
-  };
 }
 
 /** Formats dead-lettered delivery queue entries for text health output. */
@@ -327,477 +190,6 @@ export function formatConfigReloadHealthLine(summary: HealthSummary): string | n
 
 const resolveHeartbeatSummary = (cfg: OpenClawConfig, agentId: string) =>
   resolveHeartbeatSummaryForAgent(cfg, agentId);
-
-const resolveAgentOrder = (cfg: OpenClawConfig) => {
-  const defaultAgentId = resolveDefaultAgentId(cfg);
-  const entries = listAgentEntries(cfg);
-  const seen = new Set<string>();
-  const ordered: Array<{ id: string; name?: string }> = [];
-
-  for (const entry of entries) {
-    if (!entry || typeof entry !== "object") {
-      continue;
-    }
-    if (typeof entry.id !== "string" || !entry.id.trim()) {
-      continue;
-    }
-    const id = normalizeAgentId(entry.id);
-    if (!id || seen.has(id)) {
-      continue;
-    }
-    seen.add(id);
-    ordered.push({ id, name: typeof entry.name === "string" ? entry.name : undefined });
-  }
-
-  if (!seen.has(defaultAgentId)) {
-    ordered.unshift({ id: defaultAgentId });
-  }
-
-  if (ordered.length === 0) {
-    ordered.push({ id: defaultAgentId });
-  }
-
-  return { defaultAgentId, ordered };
-};
-
-const buildSessionSummary = async (storePath: string, agentId?: string) => {
-  const { listSessionEntriesReadOnly } = await import("../config/sessions/session-accessor.js");
-  const { isTransientSqliteError } = await import("../infra/unhandled-rejections.js");
-  let listed: ReturnType<typeof listSessionEntriesReadOnly>;
-  try {
-    listed = listSessionEntriesReadOnly({
-      ...(agentId ? { agentId } : {}),
-      storePath,
-    });
-  } catch (error) {
-    if (!isTransientSqliteError(error)) {
-      throw error;
-    }
-    // Health is best-effort: an empty snapshot beats failing on a transient lock.
-    listed = [];
-  }
-  const sessions = listed
-    .filter(({ sessionKey }) => sessionKey !== "global" && sessionKey !== "unknown")
-    .map(({ sessionKey, entry }) => ({ key: sessionKey, updatedAt: entry?.updatedAt ?? 0 }))
-    .toSorted((a, b) => b.updatedAt - a.updatedAt);
-  const recent = sessions.slice(0, 5).map((s) => ({
-    key: s.key,
-    updatedAt: s.updatedAt || null,
-    age: s.updatedAt ? Date.now() - s.updatedAt : null,
-  }));
-  return {
-    path: storePath,
-    count: sessions.length,
-    recent,
-  } satisfies HealthSummary["sessions"];
-};
-
-function buildPluginHealthSummary(): PluginHealthSummary | undefined {
-  const registry = getActivePluginRegistry();
-  const degradedPlugins = listActiveDegradedPlugins();
-  const unavailable = degradedPlugins
-    .map(({ pluginId, state, diagnostic }) => ({
-      id: pluginId,
-      state,
-      diagnostic: toPublicPluginVerificationDiagnostic(diagnostic),
-    }))
-    .toSorted((left, right) => left.id.localeCompare(right.id));
-  const loaded = (registry?.plugins ?? [])
-    .filter((plugin) => plugin.status === "loaded")
-    .map((plugin) => plugin.id)
-    .toSorted((left, right) => left.localeCompare(right));
-  const errors = (registry?.plugins ?? [])
-    .filter(
-      (plugin) =>
-        plugin.status === "error" &&
-        !degradedPlugins.some(
-          (degraded) =>
-            plugin.id === degraded.pluginId &&
-            plugin.failurePhase === "validation" &&
-            plugin.activationReason === `configured-unavailable: ${degraded.diagnostic.reason}` &&
-            Boolean(plugin.rootDir) &&
-            degradedPluginMatchesRoot(degraded, plugin.rootDir ?? ""),
-        ),
-    )
-    .map((plugin) => {
-      const error: PluginHealthErrorSummary = {
-        id: plugin.id,
-        origin: plugin.origin,
-        activated: plugin.activated === true,
-        error: plugin.error ?? "unknown plugin load error",
-      };
-      if (plugin.activationSource) {
-        error.activationSource = plugin.activationSource;
-      }
-      if (plugin.activationReason) {
-        error.activationReason = plugin.activationReason;
-      }
-      if (plugin.failurePhase) {
-        error.failurePhase = plugin.failurePhase;
-      }
-      return error;
-    })
-    .toSorted((left, right) => left.id.localeCompare(right.id));
-  if (loaded.length === 0 && errors.length === 0 && unavailable.length === 0) {
-    return undefined;
-  }
-  return { loaded, errors, unavailable };
-}
-
-function readBooleanField(value: unknown, key: string): boolean | undefined {
-  const record = asNullableRecord(value);
-  if (!record) {
-    return undefined;
-  }
-  return typeof record[key] === "boolean" ? record[key] : undefined;
-}
-
-const hasAccountValue = (account: unknown): boolean => account !== null && account !== undefined;
-
-function resolveProbeAccountEnabled(params: {
-  plugin: ChannelPlugin;
-  cfg: OpenClawConfig;
-  accountId: string;
-  account: unknown;
-  diagnostics: string[];
-}): boolean {
-  const fallback = readBooleanField(params.account, "enabled") ?? true;
-  try {
-    return resolveChannelAccountEnabled({
-      plugin: params.plugin,
-      account: params.account,
-      cfg: params.cfg,
-    });
-  } catch (error) {
-    params.diagnostics.push(
-      `${params.plugin.id}:${params.accountId}: failed to evaluate enabled state (${formatErrorMessage(error)}).`,
-    );
-    return fallback;
-  }
-}
-
-async function resolveProbeAccountConfigured(params: {
-  plugin: ChannelPlugin;
-  cfg: OpenClawConfig;
-  accountId: string;
-  account: unknown;
-  diagnostics: string[];
-}): Promise<boolean> {
-  const fallback = readBooleanField(params.account, "configured") ?? true;
-  try {
-    return await resolveChannelAccountConfigured({
-      plugin: params.plugin,
-      account: params.account,
-      cfg: params.cfg,
-      readAccountConfiguredField: true,
-    });
-  } catch (error) {
-    params.diagnostics.push(
-      `${params.plugin.id}:${params.accountId}: failed to evaluate configured state (${formatErrorMessage(error)}).`,
-    );
-    return fallback;
-  }
-}
-
-async function resolveHealthAccountContext(params: {
-  plugin: ChannelPlugin;
-  cfg: OpenClawConfig;
-  accountId: string;
-}): Promise<{
-  probeAccount: unknown;
-  snapshotAccount: unknown;
-  enabled: boolean;
-  configured: boolean;
-  diagnostics: string[];
-}> {
-  const diagnostics: string[] = [];
-  let account: unknown;
-  try {
-    account = params.plugin.config.resolveAccount(params.cfg, params.accountId);
-  } catch (error) {
-    diagnostics.push(
-      `${params.plugin.id}:${params.accountId}: failed to resolve account (${formatErrorMessage(error)}).`,
-    );
-  }
-  let inspectedAccount: unknown;
-  try {
-    inspectedAccount = await inspectChannelAccount(params);
-  } catch (error) {
-    diagnostics.push(
-      `${params.plugin.id}:${params.accountId}: failed to inspect account (${formatErrorMessage(error)}).`,
-    );
-  }
-
-  const probeAccount = hasAccountValue(account) ? account : inspectedAccount;
-  if (!hasAccountValue(probeAccount)) {
-    return {
-      probeAccount: {},
-      snapshotAccount: {},
-      enabled: false,
-      configured: false,
-      diagnostics,
-    };
-  }
-  const snapshotAccount = hasAccountValue(inspectedAccount) ? inspectedAccount : probeAccount;
-
-  const enabled = resolveProbeAccountEnabled({
-    plugin: params.plugin,
-    cfg: params.cfg,
-    accountId: params.accountId,
-    account: probeAccount,
-    diagnostics,
-  });
-  const configured = await resolveProbeAccountConfigured({
-    plugin: params.plugin,
-    cfg: params.cfg,
-    accountId: params.accountId,
-    account: probeAccount,
-    diagnostics,
-  });
-
-  return {
-    probeAccount,
-    snapshotAccount,
-    enabled,
-    configured,
-    diagnostics,
-  };
-}
-
-/** Builds the gateway-side health snapshot for channels, agents, plugins, and sessions. */
-export async function getHealthSnapshot(params?: {
-  timeoutMs?: number;
-  probe?: boolean;
-  includeSensitive?: boolean;
-  runtimeSnapshot?: ChannelRuntimeSnapshot;
-  eventLoop?: HealthSummary["eventLoop"];
-  configReloadHotReloadStatus?: GatewayHotReloadStatus;
-}): Promise<HealthSummary> {
-  const timeoutMs = params?.timeoutMs;
-  const cfg = await readRuntimeHealthConfig();
-  const { defaultAgentId, ordered } = resolveAgentOrder(cfg);
-  const channelBindings = buildChannelAccountBindings(cfg);
-  const sessionCache = new Map<string, HealthSummary["sessions"]>();
-  const agents: AgentHealthSummary[] = [];
-  for (const entry of ordered) {
-    const storePath = resolveStorePath(cfg.session?.store, { agentId: entry.id });
-    const sessionCacheKey = `${storePath}\0${entry.id}`;
-    const sessions =
-      sessionCache.get(sessionCacheKey) ?? (await buildSessionSummary(storePath, entry.id));
-    sessionCache.set(sessionCacheKey, sessions);
-    agents.push({
-      agentId: entry.id,
-      name: entry.name,
-      isDefault: entry.id === defaultAgentId,
-      heartbeat: resolveHeartbeatSummary(cfg, entry.id),
-      sessions,
-    });
-  }
-  const defaultAgent = agents.find((agent) => agent.isDefault) ?? agents[0];
-  const heartbeatSeconds = defaultAgent?.heartbeat.everyMs
-    ? Math.round(defaultAgent.heartbeat.everyMs / 1000)
-    : 0;
-  const sessions =
-    defaultAgent?.sessions ??
-    (await buildSessionSummary(
-      resolveStorePath(cfg.session?.store, { agentId: defaultAgentId }),
-      defaultAgentId,
-    ));
-
-  const start = Date.now();
-  const cappedTimeout = resolveTimerTimeoutMs(timeoutMs, DEFAULT_TIMEOUT_MS, 50);
-  const doProbe = params?.probe !== false;
-  const includeSensitive = params?.includeSensitive !== false;
-  const channels: Record<string, ChannelHealthSummary> = {};
-  const plugins = listReadOnlyChannelPluginsForConfig(cfg, {
-    includeSetupFallbackPlugins: false,
-  });
-  const channelOrder = plugins.map((plugin) => plugin.id);
-  const channelLabels: Record<string, string> = {};
-
-  for (const plugin of plugins) {
-    channelLabels[plugin.id] = plugin.meta.label ?? plugin.id;
-    const accountIds = plugin.config.listAccountIds(cfg);
-    const defaultAccountId = resolveChannelDefaultAccountId({
-      plugin,
-      cfg,
-      accountIds,
-    });
-    const boundAccounts = channelBindings.get(plugin.id)?.get(defaultAgentId) ?? [];
-    const preferredAccountId = resolvePreferredAccountId({
-      accountIds,
-      defaultAccountId,
-      boundAccounts,
-    });
-    const boundAccountIdsAll = Array.from(
-      new Set(Array.from(channelBindings.get(plugin.id)?.values() ?? []).flat()),
-    );
-    const accountIdsToProbe = Array.from(
-      new Set(
-        [preferredAccountId, defaultAccountId, ...accountIds, ...boundAccountIdsAll].filter(
-          (value) => value && value.trim(),
-        ),
-      ),
-    );
-    // Probe preferred/default/bound accounts first, but include all configured
-    // accounts so verbose health can explain account-specific failures.
-    debugHealth(cfg, "channel", {
-      id: plugin.id,
-      accountIds,
-      defaultAccountId,
-      boundAccounts,
-      preferredAccountId,
-      accountIdsToProbe,
-    });
-    const accountSummaries: Record<string, ChannelAccountHealthSummary> = {};
-
-    for (const accountId of accountIdsToProbe) {
-      const { probeAccount, snapshotAccount, enabled, configured, diagnostics } =
-        await resolveHealthAccountContext({
-          plugin,
-          cfg,
-          accountId,
-        });
-      if (diagnostics.length > 0) {
-        debugHealth(cfg, "account.diagnostics", { channel: plugin.id, accountId, diagnostics });
-      }
-
-      let probe: unknown;
-      let lastProbeAt: number | null = null;
-      if (enabled && configured && doProbe && plugin.status?.probeAccount) {
-        try {
-          probe = await plugin.status.probeAccount({
-            account: probeAccount,
-            timeoutMs: cappedTimeout,
-            cfg,
-          });
-          lastProbeAt = Date.now();
-        } catch (err) {
-          probe = { ok: false, error: formatErrorMessage(err) };
-          lastProbeAt = Date.now();
-        }
-      }
-
-      const probeRecord =
-        probe && typeof probe === "object" ? (probe as Record<string, unknown>) : null;
-      const bot =
-        probeRecord && typeof probeRecord.bot === "object"
-          ? (probeRecord.bot as { username?: string | null })
-          : null;
-      if (bot?.username) {
-        debugHealth(cfg, "probe.bot", { channel: plugin.id, accountId, username: bot.username });
-      }
-
-      const runtimeSnapshot =
-        params?.runtimeSnapshot?.channelAccounts[plugin.id]?.[accountId] ??
-        (accountId === defaultAccountId ? params?.runtimeSnapshot?.channels[plugin.id] : undefined);
-      const nonSensitiveProbeFailure = buildNonSensitiveProbeFailure(plugin.id, probe);
-      const snapshotProbe = includeSensitive ? probe : nonSensitiveProbeFailure;
-      const snapshot: ChannelAccountSnapshot = await buildChannelAccountSnapshotFromAccount({
-        plugin,
-        cfg,
-        accountId,
-        account: snapshotAccount,
-        runtime: runtimeSnapshot,
-        probe: snapshotProbe,
-        enabledFallback: enabled,
-        configuredFallback: configured,
-      });
-      if (lastProbeAt) {
-        snapshot.lastProbeAt = lastProbeAt;
-      }
-      const health = evaluateChannelHealth(snapshot, {
-        channelId: plugin.id,
-        now: Date.now(),
-        staleEventThresholdMs: DEFAULT_CHANNEL_STALE_EVENT_THRESHOLD_MS,
-        channelConnectGraceMs: DEFAULT_CHANNEL_CONNECT_GRACE_MS,
-      });
-      if (!health.healthy) {
-        snapshot.healthState = health.reason;
-      }
-
-      const summary = plugin.status?.buildChannelSummary
-        ? await plugin.status.buildChannelSummary({
-            account: probeAccount,
-            cfg,
-            defaultAccountId: accountId,
-            snapshot,
-          })
-        : undefined;
-      // Summary hooks overlay the safe snapshot, so reapply URL redaction after the final merge.
-      const record = redactChannelStatusSummaryBaseUrl(
-        summary && typeof summary === "object"
-          ? ({ ...snapshot, ...summary } as ChannelAccountHealthSummary)
-          : ({ ...snapshot, accountId, configured } satisfies ChannelAccountHealthSummary),
-      );
-      if (record.configured === undefined) {
-        record.configured = configured;
-      }
-      if (includeSensitive && record.probe === undefined && probe !== undefined) {
-        record.probe = probe;
-      }
-      if (!includeSensitive) {
-        const summaryProbeFailure = buildNonSensitiveProbeFailure(plugin.id, record.probe);
-        const safeProbeFailure = summaryProbeFailure ?? nonSensitiveProbeFailure;
-        if (safeProbeFailure) {
-          record.probe = safeProbeFailure;
-        } else {
-          delete record.probe;
-        }
-      }
-      if (record.lastProbeAt === undefined && lastProbeAt) {
-        record.lastProbeAt = lastProbeAt;
-      }
-      record.accountId = accountId;
-      accountSummaries[accountId] = record;
-    }
-
-    const defaultSummary =
-      accountSummaries[preferredAccountId] ??
-      accountSummaries[defaultAccountId] ??
-      accountSummaries[accountIdsToProbe[0] ?? preferredAccountId];
-    const fallbackSummary =
-      defaultSummary ??
-      accountSummaries[
-        expectDefined(Object.keys(accountSummaries)[0], "object.keys(account summaries) entry at 0")
-      ];
-    if (fallbackSummary) {
-      channels[plugin.id] = {
-        ...fallbackSummary,
-        accounts: accountSummaries,
-      } satisfies ChannelHealthSummary;
-    }
-  }
-
-  const pluginHealth = buildPluginHealthSummary();
-  const contextEngineHealth = buildContextEngineHealthSummary();
-  const deliveryQueueHealth = buildDeliveryQueueHealthSummary();
-  const summary: HealthSummary = {
-    ok: true,
-    ts: Date.now(),
-    durationMs: Date.now() - start,
-    ...(params?.eventLoop ? { eventLoop: params.eventLoop } : {}),
-    ...(pluginHealth ? { plugins: pluginHealth } : {}),
-    ...(contextEngineHealth ? { contextEngines: contextEngineHealth } : {}),
-    ...(deliveryQueueHealth ? { deliveryQueues: deliveryQueueHealth } : {}),
-    ...(params?.configReloadHotReloadStatus
-      ? { configReload: { hotReloadStatus: params.configReloadHotReloadStatus } }
-      : {}),
-    channels,
-    channelOrder,
-    channelLabels,
-    heartbeatSeconds,
-    defaultAgentId,
-    agents,
-    sessions: {
-      path: sessions.path,
-      count: sessions.count,
-      recent: sessions.recent,
-    },
-  };
-
-  return summary;
-}
 
 /** Runs the `openclaw health` command against the gateway and renders JSON or text. */
 export async function healthCommand(
@@ -1126,9 +518,3 @@ async function readBestEffortHealthConfig(): Promise<OpenClawConfig> {
   const { readBestEffortConfig } = await loadConfigRuntime();
   return await readBestEffortConfig();
 }
-
-async function readRuntimeHealthConfig(): Promise<OpenClawConfig> {
-  const { getRuntimeConfig } = await loadConfigRuntime();
-  return getRuntimeConfig();
-}
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
