+++ src/gateway/server-node-events-types.ts
@@ -2,8 +2,8 @@
 // Defines the narrowed context and event envelope for node-originated handlers.
 import type { ModelCatalogEntry } from "../agents/model-catalog.js";
 import type { CliDeps } from "../cli/deps.types.js";
-import type { HealthSummary } from "../commands/health.js";
 import type { ChatAbortControllerEntry } from "./chat-abort.js";
+import type { HealthSummary } from "./health/types.js";
 import type { ChatRunEntry, ChatRunRegistration } from "./server-chat.js";
 import type { DedupeEntry } from "./server-shared.js";
