+++ src/gateway/test-helpers.maintenance-state.ts
@@ -1,6 +1,6 @@
 // Gateway maintenance-state test helper.
 // Builds minimal timer/health/chat state for maintenance tests.
-import type { HealthSummary } from "../commands/health.js";
+import type { HealthSummary } from "./health/types.js";
 import { createChatRunState } from "./server-chat-state.js";
 
 /** Create a Gateway maintenance-state stub with configurable health/presence versions. */
